requirements.txt
pandas
numpy
python-dotenv
certifi
pymongo
scikit-learn
pymongo[srv]==3.6
pyaml
dill
mlflow